+++ addons/event_sale/static/src/js/event_configurator_test_ui.js
@@ -18,17 +18,10 @@ tour.register('event_configurator_tour', {
 }, {
     trigger: "a:contains('Add a product')"
 }, {
-    trigger: 'div[name="product_id"] input',
-    run: function (){
-        var $input = $('div[name="product_id"] input');
-        $input.click();
-        $input.val('EVENT');
-        // fake keydown to trigger search
-        var keyDownEvent = jQuery.Event("keydown");
-        keyDownEvent.which = 42;
-        $input.trigger(keyDownEvent);
-    },
-    id: 'product_selection_step'
+    trigger: 'div[name="product_id"] input, div[name="product_template_id"] input',
+    run: function (actions){
+        actions.text('EVENT');
+    }
 }, {
     trigger: 'ul.ui-autocomplete a:contains("EVENT")',
     run: 'click'
@@ -66,7 +59,7 @@ tour.register('event_configurator_tour', {
     trigger: 'td:contains("EVENT")',
     run: 'click'
 }, {
-    trigger: '.o_event_sale_js_event_configurator_edit'
+    trigger: '.o_edit_product_configuration'
 }, {
     trigger: 'div[name="event_ticket_id"] input',
     run: 'click'
