+++ addons/event_sale/views/assets.xml
@@ -1,13 +1,9 @@
 <?xml version="1.0" encoding="utf-8"?>
 <odoo>
 <template id="assets_backend" inherit_id="web.assets_backend" name="event_sale assets backend">
-    <xpath expr="link[last()]" position="after">
-        <link rel="stylesheet" type="text/scss" href="/event_sale/static/src/scss/event_configurator_widget.scss"/>
-    </xpath>
     <xpath expr="script[last()]" position="after">
         <script type="text/javascript" src="/event_sale/static/src/js/event_configurator_controller.js"></script>
         <script type="text/javascript" src="/event_sale/static/src/js/event_configurator_view.js"></script>
-        <script type="text/javascript" src="/event_sale/static/src/js/event_configurator_widget_mixin.js"></script>
         <script type="text/javascript" src="/event_sale/static/src/js/event_configurator_widget.js"></script>
         <script type="text/javascript" src="/event_sale/static/src/js/event_configurator_test_ui.js"></script>
     </xpath>
