+++ addons/event_sale/views/sale_order_views.xml
@@ -31,7 +31,6 @@
                 <field name="event_ok" invisible="1"/>
             </xpath>
             <xpath expr="//field[@name='order_line']//tree//field[@name='product_id']" position="attributes">
-                <attribute name="widget">event_configurator</attribute>
                 <attribute name="options">{'no_open': True}</attribute>
             </xpath>
             <xpath expr="//field[@name='order_line']//tree//field[@name='product_id']" position="after">
