--- addons/event_sale_product_configurator/views/assets.xml
@@ -1,9 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<odoo>
-    <template id="assets_backend" inherit_id="web.assets_backend" name="event_sale_product_configurator assets backend">
-        <xpath expr="script[last()]" position="after">
-            <script type="text/javascript" src="/event_sale_product_configurator/static/src/js/event_configurator_test_ui.js"></script>
-            <script type="text/javascript" src="/event_sale_product_configurator/static/src/js/product_configurator_widget.js"></script>
-        </xpath>
-    </template>
-</odoo>
