+++ addons/sale/static/src/scss/product_configurator.scss
@@ -218,3 +218,13 @@ label.css_attribute_color.css_not_available {
     text-decoration: line-through;
     white-space: nowrap;
 }
+
+.o_list_view {
+    .o_data_row.o_selected_row > .o_data_cell:not(.o_readonly_modifier) {
+        .o_field_widget .o_edit_product_configuration {
+            padding: 0;
+            background-color: inherit;
+            margin-left: 3px;
+        }
+    }
+}
