+++ addons/sale/views/assets.xml
@@ -9,6 +9,7 @@
             <script type="text/javascript" src="/sale/static/src/js/sale.js"></script>
             <script type="text/javascript" src="/sale/static/src/js/tour.js"></script>
             <script type="text/javascript" src="/sale/static/src/js/account_reconciliation.js"></script>
+            <script type="text/javascript" src="/sale/static/src/js/product_configurator_widget.js"/>
         </xpath>
     </template>
     <template id="assets_frontend_inherit_sale" inherit_id="web.assets_frontend" name="Sale frontend assets">
