+++ addons/sale_product_configurator/static/src/js/product_configurator_widget.js
@@ -1,138 +1,103 @@
 odoo.define('sale_product_configurator.product_configurator', function (require) {
-var relationalFields = require('web.relational_fields');
-var FieldsRegistry = require('web.field_registry');
-var core = require('web.core');
-var _t = core._t;
+var ProductConfiguratorWidget = require('sale.product_configurator');
 
 /**
- * The product configurator widget is a simple FieldMany2One that adds the capability
- * to configure a product_template_id using the product configurator wizard.
+ * Extension of the ProductConfiguratorWidget to support product configuration.
+ * It opens when a configurable product_template is set.
+ * (multiple variants, or custom attributes)
  *
- * !!! It should only be used on a product_template_id field !!!
+ * The product customization information includes :
+ * - is_configurable_product
+ * - product_template_attribute_value_ids
+ *
  */
-var ProductConfiguratorWidget = relationalFields.FieldMany2One.extend({
-    events: _.extend({}, relationalFields.FieldMany2One.prototype.events, {
-        'click .o_edit_product_configuration': '_onEditProductConfiguration'
-    }),
-
+ProductConfiguratorWidget.include({
     /**
-     * This method will check if the current product_template set on the SO line is configurable
-     * -> If so, we add a 'Edit Configuration' button next to the dropdown.
+     * Override of sale.product_configurator Hook
      *
      * @override
-     */
-    start: function () {
-        var prom = this._super.apply(this, arguments);
-
-        var $inputDropdown = this.$('.o_input_dropdown');
-
-        if (this.recordData.is_configurable_product &&
-            $inputDropdown.length !== 0 &&
-            this.$('.o_edit_product_configuration').length === 0) {
-            var $editConfigurationButton = $('<button>', {
-                type: 'button',
-                class: 'fa fa-pencil btn btn-secondary o_edit_product_configuration',
-                tabindex: '-1',
-                draggable: false,
-                'aria-label': _t('Edit Configuration'),
-                title: _t('Edit Configuration')
-            });
-
-            $inputDropdown.after($editConfigurationButton);
-        }
-        return prom;
+    */
+    _isConfigurableProduct: function() {
+        return this.recordData.is_configurable_product || this._super.apply(this, arguments);
     },
 
     /**
-     * This method is overridden to check to check if the product_template_id
-     * needs configuration or not:
+     * Set restoreProductTemplateId for further backtrack.
      *
-     * - The product_template has only one "product.product" and is not dynamic
-     *   -> Set the product_id on the SO line
-     *   -> If the product has optional products, open the configurator in 'options' mode
-     *
-     * - The product_template is configurable
-     *   -> Open the product configurator wizard and initialize it with
-     *      the provided product_template_id and its current attribute values
      * @override
-     * @param {OdooEvent} event
-     *   {boolean} event.data.preventProductIdCheck prevent the product configurator widget
-     *     from looping forever when it needs to change the 'product_template_id'
-     *
      * @private
      */
-    _onFieldChanged: function (event) {
-        var self = this;
-        self.restoreProductTemplateId = self.recordData.product_template_id;
+    _onFieldChanged: function (ev) {
+        this.restoreProductTemplateId = this.recordData.product_template_id;
 
         this._super.apply(this, arguments);
-
-        var $inputDropdown = this.$('.o_input_dropdown');
-        if (event.data.changes.product_template_id
-            && $inputDropdown.length !== 0 &&
-            this.$('.o_edit_product_configuration').length !== 0){
-            this.$('.o_edit_product_configuration').remove();
-        }
-
-        if (!event.data.changes.product_template_id
-            || event.data.preventProductIdCheck){
-            return;
-        }
-
-        var productTemplateId = event.data.changes.product_template_id.id;
-        if (productTemplateId){
-            this._rpc({
-                model: 'product.template',
-                method: 'get_single_product_variant',
-                args: [
-                    productTemplateId
-                ]
-            }).then(function (result){
-                if (result){
-                    self.trigger_up('field_changed', {
-                        dataPointID: event.data.dataPointID,
-                        changes: {
-                            product_id: {id: result.product_id},
-                            product_custom_attribute_value_ids: {
-                                operation: 'DELETE_ALL'
-                            }
-                        },
-                        onSuccess: function () {
-                            if (result.has_optional_products) {
-                                self._openProductConfigurator({
-                                        configuratorMode: 'options',
-                                        default_pricelist_id: self._getPricelistId(),
-                                        default_product_template_id: productTemplateId
-                                    },
-                                    event.data.dataPointID
-                                );
-                            }
-                        }
-                    });
-
-                    self._onSimpleProductFound(result.product_id, event.data.dataPointID);
-                } else {
-                    self._openProductConfigurator({
-                            configuratorMode: 'add',
-                            default_pricelist_id: self._getPricelistId(),
-                            default_product_template_id: productTemplateId
-                        },
-                        event.data.dataPointID
-                    );
-                }
-            });
-        }
     },
 
     /**
-     * Hooking point for other modules
-     *
-     * @param {integer} productId
-     * @param {string} dataPointID
-     *
-     * @private
-     */
-    _onSimpleProductFound: function (productId, dataPointID) {},
+    * This method is overridden to check if the product_template_id
+    * needs configuration or not:
+    *
+    * - The product_template has only one "product.product" and is not dynamic
+    *   -> Set the product_id on the SO line
+    *   -> If the product has optional products, open the configurator in 'options' mode
+    *
+    * - The product_template is configurable
+    *   -> Open the product configurator wizard and initialize it with
+    *      the provided product_template_id and its current attribute values
+    *
+    * @override
+    * @private
+    */
+    _onTemplateChange: function (productTemplateId, dataPointId) {
+        var self = this;
+
+        return this._rpc({
+            model: 'product.template',
+            method: 'get_single_product_variant',
+            args: [
+                productTemplateId
+            ]
+        }).then(function (result){
+            if (result){
+                self.trigger_up('field_changed', {
+                    dataPointID: dataPointId,
+                    changes: {
+                        product_id: {id: result.product_id},
+                        product_custom_attribute_value_ids: {
+                            operation: 'DELETE_ALL'
+                        }
+                    },
+                    onSuccess: function () {
+                        if (result.has_optional_products) {
+                            // VFE WARNING optional products and configurable product_id don't go well
+                            // together for the moment...
+                            self._openOptionsConfigurator(productTemplateId, dataPointId);
+                        }
+                    }
+                });
+            } else {
+                self._openProductConfigurator({
+                        configuratorMode: 'add',
+                        default_pricelist_id: self._getPricelistId(),
+                        default_product_template_id: productTemplateId
+                    },
+                    dataPointId
+                );
+            }
+            return Promise.resolve(true);
+            // always returns true for the moment because no other configurator exists.
+        });
+    },
+
+    _openOptionsConfigurator: function (productTemplateId, dataPointId) {
+        this._openProductConfigurator({
+                configuratorMode: 'options',
+                default_pricelist_id: this._getPricelistId(),
+                default_product_template_id: productTemplateId
+            },
+            dataPointId
+        );
+    },
 
     /**
      * Opens the product configurator to allow configuring the product template
@@ -150,7 +115,7 @@ var ProductConfiguratorWidget = relationalFields.FieldMany2One.extend({
      * the product_template is reset to its previous value if any.
      *
      * @param {Object} data various "default_" values
-     *  {string} data.configuratorMode 'add' or 'edit'.
+     *  {string} data.configuratorMode 'add' or 'edit' or 'options'.
      * @param {string} dataPointId
      *
      * @private
@@ -161,6 +126,8 @@ var ProductConfiguratorWidget = relationalFields.FieldMany2One.extend({
             additional_context: data,
             on_close: function (result) {
                 if (result && result !== 'special'){
+                    // TODO if mode == 'options'
+                    // if no options chosen, don't trigger events...
                     self._addProducts(result, dataPointId);
                 } else {
                     if (self.restoreProductTemplateId) {
@@ -186,6 +153,13 @@ var ProductConfiguratorWidget = relationalFields.FieldMany2One.extend({
      * @private
      */
     _onEditProductConfiguration: function () {
+        if (!this.recordData.is_configurable_product) {
+            // if line should be edited by another configurator
+            // or simply inline.
+            this._super.apply(this, arguments);
+            return;
+        }
+        // If line has been set up through the product_configurator:
         this._openProductConfigurator({
                 configuratorMode: 'edit',
                 default_product_template_id: this.recordData.product_template_id.data.id,
@@ -230,16 +204,22 @@ var ProductConfiguratorWidget = relationalFields.FieldMany2One.extend({
             preventProductIdCheck: true,
             changes: this._getMainProductChanges(result.mainProduct),
             onSuccess: function () {
-                if (result.options) {
-                    var parentList = self.getParent();
+                var parentList = self.getParent();
+                if (result.options && result.options.length !== 0) {
                     self.trigger_up('add_record', {
                         context: self._productsToRecords(result.options),
                         forceEditable: 'bottom',
                         allowWarning: true,
-                        onSuccess: function (){
+                        onSuccess: function () {
+                            // Leave edit mode of the last line added.
                             parentList.unselectRow();
                         }
                     });
+
+                } else if (!self._isConfigurableLine()) {
+                    // Leave edit mode of current line if no other wizard
+                    // will be opened by the field_changed event.
+                    parentList.unselectRow();
                 }
             }
         });
@@ -370,64 +350,8 @@ var ProductConfiguratorWidget = relationalFields.FieldMany2One.extend({
         });
 
         return records;
-    },
-
-    /**
-     * Will convert the values contained in the recordData parameter to
-     * a list of '4' operations that can be passed as a 'default_' parameter.
-     *
-     * @param {Object} recordData
-     *
-     * @private
-     */
-    _convertFromMany2Many: function (recordData) {
-        if (recordData) {
-            var convertedValues = [];
-            _.each(recordData.res_ids, function (res_id) {
-                convertedValues.push([4, parseInt(res_id)]);
-            });
-
-            return convertedValues;
-        }
-
-        return null;
-    },
-
-    /**
-     * Will convert the values contained in the recordData parameter to
-     * a list of '0' or '4' operations (based on wether the record is already persisted or not)
-     * that can be passed as a 'default_' parameter.
-     *
-     * @param {Object} recordData
-     *
-     * @private
-     */
-    _convertFromOne2Many: function (recordData) {
-        if (recordData) {
-            var convertedValues = [];
-            _.each(recordData.res_ids, function (res_id) {
-                if (isNaN(res_id)){
-                    _.each(recordData.data, function (record) {
-                        if (record.ref === res_id) {
-                            convertedValues.push([0, 0, {
-                                attribute_value_id: record.data.attribute_value_id.data.id,
-                                custom_value: record.data.custom_value
-                            }]);
-                        }
-                    });
-                } else {
-                    convertedValues.push([4, res_id]);
-                }
-            });
-
-            return convertedValues;
-        }
-
-        return null;
     }
 });
-
-FieldsRegistry.add('product_configurator', ProductConfiguratorWidget);
 
 return ProductConfiguratorWidget;
