--- addons/sale_product_configurator/static/src/scss/product_configurator_widget.scss
@@ -1,9 +1,0 @@
-.o_list_view {
-    .o_data_row.o_selected_row > .o_data_cell:not(.o_readonly_modifier) {
-        .o_field_widget .o_edit_product_configuration {
-            padding: 0;
-            background-color: inherit;
-            margin-left: 3px;
-        }
-    }
-}
