+++ addons/sale_product_configurator/views/assets.xml
@@ -1,9 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <odoo>
     <template id="assets_backend" inherit_id="web.assets_backend" name="sale_product_configurator assets backend">
-        <xpath expr="link[last()]" position="after">
-            <link rel="stylesheet" type="text/scss" href="/sale_product_configurator/static/src/scss/product_configurator_widget.scss"/>
-        </xpath>
         <xpath expr="script[last()]" position="after">
             <script type="text/javascript" src="/sale_product_configurator/static/src/js/tour.js"></script>
             <script type="text/javascript" src="/sale/static/src/js/variant_mixin.js"></script>
