+++ addons/sale_product_configurator/views/sale_views.xml
@@ -20,26 +20,10 @@
         <field name="model">sale.order</field>
         <field name="inherit_id" ref="sale.view_order_form"/>
         <field name="arch" type="xml">
-            <xpath expr="//tree/field[@name='product_id']" position="after">
-                <field
-                    name="product_template_id"
-                    string="Product"
-                    options="{'no_open': True}"
-                    attrs="{
-                        'readonly': [('product_updatable', '=', False)],
-                        'required': [('display_type', '=', False)],
-                    }"
-                    context="{
-                        'partner_id': parent.partner_id,
-                        'quantity': product_uom_qty,
-                        'pricelist': parent.pricelist_id,
-                        'uom':product_uom,
-                        'company_id': parent.company_id,
-                        'default_lst_price': price_unit,
-                        'default_description_sale': name
-                    }"
-                    widget="product_configurator"
-                />
+            <xpath expr="//tree/field[@name='product_template_id']" position="attributes">
+                <attribute name="invisible">0</attribute>
+            </xpath>
+            <xpath expr="//tree/field[@name='product_template_id']" position="after">
                 <field name="product_template_attribute_value_ids" invisible="1" />
                 <field name="product_custom_attribute_value_ids" invisible="1" >
                     <tree>
